The run directory will only be deleted by the when running in a different process if given the delete directory command.

+++ Code/XTMF.Run/Program.cs
@@ -188,7 +188,7 @@ namespace XTMF.Run
             {
                 try
                 {
-                    if (Directory.Exists(runDirectory))
+                    if (deleteDirectory && Directory.Exists(runDirectory))
                     {
                         Directory.Delete(runDirectory, true);
                     }
